extracting custom logger into a separate module
